+++ pkg/ffuf/progress.go
@@ -8,6 +8,7 @@ type Progress struct {
 	StartedAt  time.Time
 	ReqCount   int
 	ReqTotal   int
+	ReqSec     int64
 	QueuePos   int
 	QueueTotal int
 	ErrorCount int